+++ R/police-speeding-band.R
@@ -70,9 +70,9 @@
 #' July 2009.
 #'
 #' @source
-#' \url{http://www.police.govt.nz/about-us/publication/road-policing-driver-offence-data-january-2009-december-2017}
+#' \url{http://www.police.govt.nz/about-us/publication/road-policing-driver-offence-data-january-2009-march-2018}
 #'
 #' @seealso \code{\link{driving_offences}}, \code{\link{excess}},
 #' \code{\link{fleeing_area}}, \code{\link{fleeing_district}},
-#' \code{\link{police_speeding}}
+#' \code{\link{police_speeding}}, \code{\link{static_camera}}
 "police_speeding_band"