+++ data-raw/fleeing.R
@@ -1,5 +1,5 @@
-url_source <- "http://www.police.govt.nz/sites/default/files/publications/road-policing-driver-offence-data-1jan2009-31dec2017.xlsx"
-path_source <-  "./inst/extdata/road-policing-driver-offence-data-1jan2009-31dec2017.xlsx"
+url_source <- "http://www.police.govt.nz/sites/default/files/publications/road-policing-driver-offence-data-1jan2009-31mar2018.xlsx"
+path_source <-  "./inst/extdata/road-policing-driver-offence-data-1jan2009-31mar2018.xlsx"
 
 library(tidyxl)
 library(unpivotr)
@@ -12,108 +12,56 @@ formats <- xlsx_formats(path_source)
 font_rgb <- formats$local$font$color$rgb
 font_size <- formats$local$font$size
 
-fleeing <- filter(cells, sheet == "Fleeing Drivers")
+fleeing <- dplyr::filter(cells, sheet == "Fleeing Drivers")
 
-# The first table on the sheet "Fleeing Drivers" is similar to the first
-# tranche, except that "Sum:" is in the first, not the second, column, and there
-# is no 'area' header.
+corners <-
+  fleeing %>%
+  dplyr::filter(font_rgb[local_format_id] == "FFFF0000",
+                font_size[local_format_id] == 10,
+                !is_blank,
+                !is.na(character))
 
-tidydistrict <- function(.anchor, cells) {
-  districts <-
-    .anchor %>%
-    offset_S(cells, 3) %>%
-    extend_S(cells, boundary = ~ character == "Sum:") %>%
-    select(row, col, district = character)
-  .months <-
-    .anchor %>%
-    offset_S(cells, 2) %>%
-    offset_E(cells, 2) %>%
-    extend_E(cells, boundary = ~ is.na(character)) %>%
-    filter(character != "Total") %>%
-    select(row, col, month = character)
-  .years <-
-    .months %>%
-    offset_N(cells, 1) %>%
-    filter(!is_blank,
-           col %in% .months$col) %>%
-    select(row, col, year = numeric)
-  .data <-
-    cells %>%
-    filter(row %in% districts$row, col %in% .months$col, !is.na(numeric)) %>%
-    select(row, col, value = numeric)
-  .data %>%
-    NNW(.anchor) %>%
-    W(districts) %>%
-    N(.months) %>%
-    NNW(.years) %>%
-    select(-row, -col)
-}
-
-get_series <- function(sheet) {
-  sheet %>%
-    filter(font_rgb[local_format_id] == "FFFF0000", # Red
-           font_size[local_format_id] == 10,
-           !is.na(character)) %>%
-    select(row, col, series = character) %>%
-    split(.$row)
-}
+partitions <-
+  fleeing %>%
+  dplyr::filter(!is_blank,
+                font_size[local_format_id] == 10) %>%
+  select(row, col, data_type, character, numeric) %>%
+  partition(corners) %>%
+  mutate(series = corners$character)
 
 fleeing_district_tidied <-
-  fleeing %>%
-  get_series %>%
-  .[[1]] %>%
-  tidydistrict(fleeing) %>%
+  partitions %>%
+  dplyr::filter(str_detect(series, regex("district", ignore_case = TRUE))) %>%
+  mutate(cells = map(cells,
+                     ~ .x %>%
+                       behead("NNW", "series") %>%
+                       behead("NNW", "year") %>%
+                       behead("N", "month") %>%
+                       behead("W", "district"))) %>%
+  unnest() %>%
+  dplyr::filter(month != "Total",
+                district != "Sum:") %>%
   mutate(month = ymd(paste(year, month, 1))) %>%
-  select(-year, -series)
-
-# The second table on the sheet "Fleeing Drivers" is similar to the first
-# tranche, except that "Sum:" is in the first, not the second, column.
-
-tidyarea <- function(.anchor, cells) {
-  districts <-
-    .anchor %>%
-    offset_S(cells, 3) %>%
-    extend_S(cells, boundary = ~ character == "Sum:") %>%
-    select(row, col, district = character)
-  areas <-
-    .anchor %>%
-    offset_S(cells, 3) %>%
-    offset_E(cells, 1) %>%
-    extend_S(cells, boundary = ~ is.na(character)) %>%
-    select(row, col, area = character)
-  .months <-
-    .anchor %>%
-    offset_S(cells, 2) %>%
-    offset_E(cells, 2) %>%
-    extend_E(cells, boundary = ~ is.na(character)) %>%
-    filter(character != "Total") %>%
-    select(row, col, month = character)
-  .years <-
-    .months %>%
-    offset_N(cells, 1) %>%
-    filter(!is_blank,
-           col %in% .months$col) %>%
-    select(row, col, year = numeric)
-  .data <-
-    cells %>%
-    filter(row %in% areas$row, col %in% .months$col, !is.na(numeric)) %>%
-    select(row, col, value = numeric)
-  .data %>%
-    NNW(.anchor) %>%
-    W(districts) %>%
-    W(areas) %>%
-    N(.months) %>%
-    NNW(.years) %>%
-    select(-row, -col)
-}
+  select(value = numeric, series, district, month)
 
 fleeing_area_tidied <-
-  fleeing %>%
-  get_series %>%
-  .[[2]] %>%
-  tidyarea(fleeing) %>%
+  partitions %>%
+  dplyr::filter(str_detect(series, regex("area", ignore_case = TRUE))) %>%
+  mutate(cells = map(cells,
+                     ~ .x %>%
+                       behead("NNW", "series") %>%
+                       behead("NNW", "year") %>%
+                       behead("N", "month") %>%
+                       behead("W", "district") %>%
+                       behead("W", "area"))) %>%
+  unnest() %>%
+  dplyr::filter(month != "Total",
+                district != "Sum:") %>%
   mutate(month = ymd(paste(year, month, 1))) %>%
-  select(-year, -series)
+  select(value = numeric, series, district, area, month)
 
 write.csv(fleeing_district_tidied, row.names = FALSE, quote = FALSE,
           file=gzfile("./inst/extdata/fleeing-drivers-district.csv.gz"))
+
+write.csv(fleeing_area_tidied, row.names = FALSE, quote = FALSE,
+          file=gzfile("./inst/extdata/fleeing-drivers-area.csv.gz"))