+++ data-raw/general-sheets.R
@@ -1,5 +1,5 @@
-url_source <- "http://www.police.govt.nz/sites/default/files/publications/road-policing-driver-offence-data-1jan2009-31dec2017.xlsx"
-path_source <-  "./inst/extdata/road-policing-driver-offence-data-1jan2009-31dec2017.xlsx"
+url_source <- "http://www.police.govt.nz/sites/default/files/publications/road-policing-driver-offence-data-1jan2009-31mar2018.xlsx"
+path_source <-  "./inst/extdata/road-policing-driver-offence-data-1jan2009-31mar2018.xlsx"
 
 library(tidyxl)
 library(unpivotr)
@@ -12,97 +12,112 @@ formats <- xlsx_formats(path_source)
 font_rgb <- formats$local$font$color$rgb
 font_size <- formats$local$font$size
 
-tidy <- function(.anchor, cells) {
-  districts <-
-    .anchor %>%
-    offset_S(cells, 3) %>%
-    extend_S(cells, boundary = ~ is.na(character)) %>%
-    select(row, col, district = character)
-  areas <-
-    .anchor %>%
-    offset_S(cells, 3) %>%
-    offset_E(cells, 1) %>%
-    extend_S(cells, boundary = ~ character == "Sum:") %>%
-    select(row, col, area = character)
-  .months <-
-    .anchor %>%
-    offset_S(cells, 2) %>%
-    offset_E(cells, 2) %>%
-    extend_E(cells, boundary = ~ is.na(character)) %>%
-    filter(character != "Total") %>%
-    select(row, col, month = character)
-  .years <-
-    .months %>%
-    offset_N(cells, 1) %>%
-    filter(!is_blank,
-           col %in% .months$col) %>%
-    select(row, col, year = numeric)
-  .data <-
-    cells %>%
-    filter(row %in% areas$row, col %in% .months$col, !is.na(numeric)) %>%
-    select(row, col, value = numeric)
-  .data %>%
-    NNW(.anchor) %>%
-    W(districts) %>%
-    W(areas) %>%
-    N(.months) %>%
-    NNW(.years) %>%
-    select(-row, -col)
-}
+general_cells <-
+  cells %>%
+  dplyr::filter(!is_blank,
+                sheet %in% c("Red Light",
+                             "Restraints",
+                             "Alcohol & Drugs",
+                             "Mobile phone",
+                             "Mobile Speed Camera",
+                             "Officer-issued Speed")) %>%
+  select(sheet, row, col, data_type, numeric, character, local_format_id)
 
-get_series <- function(sheet) {
-  sheet %>%
-    filter(font_rgb[local_format_id] == "FFFF0000", # Red
-           font_size[local_format_id] == 10,
-           !is.na(character)) %>%
-    select(row, col, series = character) %>%
-    split(.$row)
-}
-
-tidy_sheets <- function(sheet) {
-  # The main series are labelled in red, font-size 10
-  sheet %>%
-    get_series %>%
-    map_df(tidy, cells = sheet)
-}
+corners <-
+  general_cells %>%
+  dplyr::filter(font_rgb[local_format_id] == "FFFF0000",
+                font_size[local_format_id] == 10) %>%
+  select(sheet, row, col, character) %>%
+  nest(-sheet)
 
 general <-
-  cells %>%
+  general_cells %>%
   nest(-sheet) %>%
-  filter(sheet %in% c("Red Light"
-                      , "Restraints"
-                      , "Alcohol & Drugs"
-                      , "Mobile phone"
-                      , "Camera-issued Speed"
-                      , "Vehicles past cameras"
-                      , "Officer-issued Speed")) %>%
-  mutate(data = map(data, tidy_sheets)) %>%
+  dplyr::mutate(data = map2(data, corners$data, partition)) %>%
   unnest() %>%
-  rename(category = sheet) %>%
+  mutate(cells = map(cells,
+                     ~ .x %>%
+                       behead("NNW", "series") %>%
+                       behead("NNW", "year") %>%
+                       behead("N", "month") %>%
+                       behead("W", "district") %>%
+                       behead("W", "area") %>%
+                       mutate(year = as.character(year)))) %>%
+  unnest() %>%
+  dplyr::filter(month != "Total", area != "Sum:", !is.na(numeric)) %>%
   mutate(month = ymd(paste(year, month, 1))) %>%
-  select(-year)
+  select(category = sheet, value = numeric, series, district, area, month)
 
-police_speeding <- filter(cells, sheet == "Police Speeding")
-police_speeding_tidied <-
-  police_speeding %>%
-  get_series %>%
-  map_df(tidy, cells = police_speeding)
+static_camera_cells <-
+  cells %>%
+  dplyr::filter(!is_blank, sheet == "Static Speed Camera") %>%
+  select(row, col, data_type, numeric, character, local_format_id)
+
+static_camera_corners <-
+  static_camera_cells %>%
+  dplyr::filter(font_rgb[local_format_id] == "FFFF0000",
+                font_size[local_format_id] == 10) %>%
+  select(row, col, character)
+
+static_camera <-
+  static_camera_cells %>%
+  partition(static_camera_corners) %>%
+  mutate(cells = map(cells,
+                     ~ .x %>%
+                       behead("NNW", "series") %>%
+                       behead("NNW", "year") %>%
+                       behead("N", "month") %>%
+                       behead("W", "district") %>%
+                       behead("W", "area") %>%
+                       behead("W", "site", formatters = list(character = str_trim)) %>%
+                       behead("W", "road", formatters = list(character = str_trim)) %>%
+                       behead("W", "section"))) %>%
+  unnest() %>%
+  dplyr::filter(month != "Total", section != "Sum:") %>%
+  mutate(month = ymd(paste(year, month, 1))) %>%
+  select(value = numeric, series, district, area, site, road, section, month)
+
+police_speeding_cells <-
+  cells %>%
+  dplyr::filter(!is_blank, sheet == "Police Speeding") %>%
+  select(row, col, data_type, numeric, character, local_format_id)
+
+police_speeding_corners <-
+  police_speeding_cells %>%
+  dplyr::filter(font_rgb[local_format_id] == "FFFF0000",
+                font_size[local_format_id] == 10) %>%
+  select(row, col, character)
+
+police_speeding <-
+  police_speeding_cells %>%
+  partition(police_speeding_corners) %>%
+  mutate(cells = map(cells,
+                     ~ .x %>%
+                       behead("NNW", "series") %>%
+                       behead("NNW", "year") %>%
+                       behead("N", "month/speed") %>%
+                       behead("W", "district") %>%
+                       behead("W", "area"))) %>%
+  unnest() %>%
+  dplyr::filter(`month/speed` != "Total", area != "Sum:")
 
 police_speeding_band <-
-  police_speeding_tidied %>%
-  filter(series == "Police vehicle speed detections by calendar year* and speed band") %>%
-  rename(speed = month) %>%
-  mutate(year = ymd(paste(year, 1, 1)))
+  police_speeding %>%
+  dplyr::filter(series == "Police vehicle speed detections by calendar year* and speed band") %>%
+  mutate(speed = `month/speed`, year = ymd(paste(year, 1, 1))) %>%
+  select(value = numeric, series, district, area, speed, year)
 
 police_speeding_tidied <-
-  filter(police_speeding_tidied,
-        series != "Police vehicle speed detections by calendar year* and speed band") %>%
-  mutate(month = ymd(paste(year, month, 1))) %>%
-  select(-year)
+  police_speeding %>%
+  dplyr::filter(series != "Police vehicle speed detections by calendar year* and speed band") %>%
+  mutate(month = ymd(paste(year, `month/speed`, 1))) %>%
+  select(value = numeric, series, district, area, month)
 
 write.csv(general, row.names = FALSE, quote = FALSE,
           file=gzfile("./inst/extdata/general-offences.csv.gz"))
-write.csv(police_speeding_tidied, row.names = FALSE, quote = FALSE,
+write.csv(static_camera, row.names = FALSE, quote = FALSE,
+          file=gzfile("./inst/extdata/static-camera.csv.gz"))
+write.csv(police_speeding, row.names = FALSE, quote = FALSE,
           file=gzfile("./inst/extdata/police_speeding.csv.gz"))
 write.csv(police_speeding_band, row.names = FALSE, quote = FALSE,
           file=gzfile("./inst/extdata/police-speeding-band.csv.gz"))